Cambio tipo dell'ID da int a long

+++ src/main/java/model/DAOInterface.java
@@ -5,7 +5,7 @@ import java.sql.SQLException;
 import java.util.Collection;
 
 public interface DAOInterface<T, K> {
-    public T doRetrieveByKey(int code) throws SQLException;
+    public T doRetrieveByKey(long code) throws SQLException;
     public Collection<T> doRetrieveAll(String order) throws SQLException;
     public void doSave(T product) throws SQLException;
     void doUpdate(T product) throws SQLException;
