+++ src/main/java/model/utente/UtenteDAO.java
@@ -15,10 +15,10 @@ public class UtenteDAO extends AbstractDAO implements DAOInterface<UtenteBean, L
     }
 
     @Override
-    public UtenteBean doRetrieveByKey(int id) throws SQLException {
+    public UtenteBean doRetrieveByKey(long id) throws SQLException {
         String query = "SELECT * FROM Utente WHERE id = ?";
         try (PreparedStatement statement = connection.prepareStatement(query)) {
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
                     return getUtente(resultSet);
